backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
psycopg2-binary==2.9.9
marshmallow==3.21.1
marshmallow-sqlalchemy==1.1.0
python-dotenv==1.0.1
gunicorn==21.2.0
